+++ OneBlink.SDK/FormsAppsClient.cs
@@ -1,7 +1,5 @@
 using System;
-using System.IdentityModel.Tokens.Jwt;
 using System.Threading.Tasks;
-using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
 using OneBlink.SDK.Model;
 
@@ -22,18 +20,14 @@ namespace OneBlink.SDK
 
         public async Task<JWTPayload> VerifyJWT(string token)
         {
+            await this.GetMyFormsApp<FormsAppBase>(token);
+
             if (token.Contains("Bearer "))
             {
                 token = token.Split(' ')[1];
             }
-            JwtSecurityToken jwt = new JwtSecurityToken(token);
-            string iss = this.oneBlinkApiClient.tenant.jwtIssuer;
-            JsonWebKey jwk = await Token.GetJsonWebKey(
-                iss: iss,
-                kid: jwt.Header.Kid
-            );
-            string verifiedToken = Token.VerifyJWT(token, jwk, iss);
-            RawJWTPayload rawJWTPayload = JsonConvert.DeserializeObject<RawJWTPayload>(verifiedToken);
+            string decodedToken = Token.DecodeJWT(token);
+            RawJWTPayload rawJWTPayload = JsonConvert.DeserializeObject<RawJWTPayload>(decodedToken);
             if (String.IsNullOrEmpty(rawJWTPayload.sub))
             {
                 return null;